+++ src/pages/accounts.jsx
@@ -1,53 +1,7 @@
 import '../style.scss';
 import ProgressSidebar from '../components/progressSidebar';
-import * as _ from 'lodash';
 import fetch from 'isomorphic-unfetch';
-
-const ConfigPropertyInput = props => (
-  <div className='config-property-input'>
-    <code>{props.displayName}</code>
-    <input
-      type='text'
-      id={props.propertyId}
-      name={props.propertyId}
-      defaultValue={props.config[props.propertyId]}
-      onChange={props.onChange}
-    />
-  </div>
-);
-
-const ConfigPropertyInputGroup = props => {
-  const handleInputChange = e => {
-    const newConfig = {
-      ...props.config,
-      [e.currentTarget.id]: e.currentTarget.value
-    };
-
-    console.log('OLD', props.config);
-    console.log('NEW', newConfig);
-    fetch('http://localhost:3000/config', {
-      method: 'PUT',
-      body: JSON.stringify(newConfig),
-      headers: {
-        'Content-Type': 'application/json'
-      }
-    });
-  };
-
-  const configPropertyInputs = _.map(props.configProperties, property => {
-    return (
-      <ConfigPropertyInput
-        config={props.config}
-        displayName={property.displayName}
-        propertyId={property.propertyId}
-        onChange={handleInputChange}
-        key={property.propertyId}
-      />
-    );
-  });
-
-  return <div className='config-container'>{configPropertyInputs}</div>;
-};
+import ConfigPropertyInputGroup from '../components/configPropertyInputGroup';
 
 const Accounts = props => {
   const configProperties = [
@@ -55,11 +9,10 @@ const Accounts = props => {
     { displayName: 'public_key', propertyId: 'PLAID_PUBLIC_KEY' },
     { displayName: 'development_secret', propertyId: 'PLAID_SECRET' }
   ];
-  console.log(props.config);
 
   return (
     <div className='wrapper'>
-      <ProgressSidebar tab='accounts' />
+      <ProgressSidebar tab='accounts' config={props.config} />
       <div className='container'>
         <div className='inner-container'>
           <h1>Plaid Setup</h1>
