+++ src/pages/index.jsx
@@ -1,16 +1,25 @@
 import '../style.scss';
 import Link from 'next/link';
 import ProgressSidebar from '../components/progressSidebar';
+import fetch from 'isomorphic-unfetch';
 
-const Welcome = () => (
+const Welcome = props => (
   <div className='wrapper'>
-    <ProgressSidebar tab='' />
+    <ProgressSidebar tab='' config={props.config} />
     <div className='container'>
-      <img style={{width: "120px"}} src="/static/icon.png"/>
+      <img style={{ width: '120px' }} src='/static/icon.png' />
       <h1>Welcome to Mintable</h1>
-      <Link href="/accounts"><button>Get Started</button></Link>
+      <Link href='/accounts'>
+        <button>Get Started</button>
+      </Link>
     </div>
   </div>
 );
+
+Welcome.getInitialProps = async function() {
+  const res = await fetch('http://localhost:3000/config');
+
+  return { config: await res.json() };
+};
 
 export default Welcome;
