+++ src/style.scss
@@ -35,7 +35,7 @@ html,
 }
 
 .inner-container {
-  max-width: 800px;
+  max-width: 1000px;
 }
 
 .progress-sidebar {
